combine the constant visitor
