+++ ExpressionToWhereClause/Constant/ConstantExpressionVisitor.cs
@@ -5,13 +5,13 @@ using System.Text;
 
 namespace ExpressionToWhereClause
 {
-    internal class ConstantExpressionVisitor: ExpressionVisitor
-    {
-        public object Value { get; set; }
-        protected override Expression VisitConstant(ConstantExpression node)
-        {
-            Value = node.Value;
-            return node;
-        }
-    }
+    //internal class ConstantExpressionVisitor: ExpressionVisitor
+    //{
+    //    public object Value { get; set; }
+    //    protected override Expression VisitConstant(ConstantExpression node)
+    //    {
+    //        Value = node.Value;
+    //        return node;
+    //    }
+    //}
 }
